refactor(server): move share web routes into package
