+++ packages/doc-server/src/share.ts
@@ -1,5 +1,5 @@
 import type { Router } from 'express';
-import { shareWebRoutes } from '../../../server/share-web-routes.js';
+import { shareWebRoutes } from './share-web-routes.js';
 
 export function createShareRouter(): Router {
   return shareWebRoutes;
